Migrate upsert_feed() and upsert_display().
