+++ model/utils/custom.py
@@ -4,57 +4,6 @@ from sqlalchemy import sql
 
 from .. import get_connection, get_table
 
-# Feed.
-def upsert_feed(feed_id, title, link, lang):
-    feeds = get_table('Feeds')
-
-    if feed_id:
-        q = feeds.update().values(
-                Title = title,
-                Link = link,
-                Language = lang.name
-        ).where(feeds.c.FeedID == feed_id)
-    else:
-        q = feeds.insert().values(
-                Title = title,
-                Link = link,
-                Language = lang.name
-        )
-        q = q.prefix_with("OR IGNORE", dialect='sqlite')
-    with get_connection() as conn:
-        conn.execute(q)
-
-    if not feed_id:
-        q = sql.select([
-                feeds.c.FeedID
-        ]).where(sql.and_(
-                feeds.c.Title == title,
-                feeds.c.Link == link,
-                feeds.c.Language == lang.name
-        ))
-        with get_connection() as conn:
-            res = conn.execute(q).fetchone()
-        return res['FeedID']
-
-def upsert_display(user_id, feed_ids, disp):
-    display = get_table('Display')
-
-    if disp == 0:
-        q = display.delete().where(sql.and_(
-                display.c.UserID == user_id,
-                display.c.FeedID.in_(feed_ids)
-        ))
-        with get_connection() as conn:
-            conn.execute(q)
-    else:
-        row_keys = ['UserID', 'FeedID']
-        rows = [dict(zip(row_keys, (user_id, e))) for e in feed_ids]
-
-        q = display.insert()
-        q = q.prefix_with("OR IGNORE", dialect='sqlite')
-        with get_connection() as conn:
-            conn.execute(q, rows)
-
 def delete_feeds(feed_ids):
     feeds = get_table('Feeds')
