+++ view/feed/db.py
@@ -30,3 +30,53 @@ def feed_tags(
     with model.get_session() as session:
         return [e[0] for e in session.execute(q)]
 
+def upsert_feed(feed_id, title, link, lang):
+    feeds = model.get_table('Feeds')
+
+    if feed_id:
+        q = feeds.update().values(
+                Title = title,
+                Link = link,
+                Language = lang.name
+        ).where(feeds.c.FeedID == feed_id)
+    else:
+        q = feeds.insert().values(
+                Title = title,
+                Link = link,
+                Language = lang.name
+        )
+        q = q.prefix_with("OR IGNORE", dialect='sqlite')
+    with model.get_connection() as conn:
+        conn.execute(q)
+
+    if not feed_id:
+        q = sqla.select([
+                feeds.c.FeedID
+        ]).where(sqla.and_(
+                feeds.c.Title == title,
+                feeds.c.Link == link,
+                feeds.c.Language == lang.name
+        ))
+        with model.get_connection() as conn:
+            res = conn.execute(q).fetchone()
+        return res['FeedID']
+
+def upsert_display(user_id, feed_ids, disp):
+    display = model.get_table('Display')
+
+    if disp == 0:
+        q = display.delete().where(sqla.and_(
+                display.c.UserID == user_id,
+                display.c.FeedID.in_(feed_ids)
+        ))
+        with model.get_connection() as conn:
+            conn.execute(q)
+    else:
+        row_keys = ['UserID', 'FeedID']
+        rows = [dict(zip(row_keys, (user_id, e))) for e in feed_ids]
+
+        q = display.insert()
+        q = q.prefix_with("OR IGNORE", dialect='sqlite')
+        with model.get_connection() as conn:
+            conn.execute(q, rows)
+
